requirements.txt
certifi==2019.6.16
ForgeryPy3==0.3.1
wincertstore==0.2
